availableMoves
available moves for selected checker piece are now calculated and a
highlight effect is placed on the squares that it can be moved to.
Going to begin working on moving the pieces to available locations now.
Need to adjust some of the code in the availableMoves and
drawAvailableMoves functions to include spaces that can be jumped to
over enemy pieces.